Refactor main plugin file.
The instantiation of main and logging class is kept in the global scope, but the injection of the logging class into the main class, and the calling of the main class run() method that hooks in all of the other actions and filters, is now done within a function hooked to the init action.

The loading of the plugin text domain / localization is also improved, and hooked into the init action as well (but in a different function).